api/requirements.txt
django==2.2.*
djangorestframework
django-cors-headers
